Added Run Test (#18)
Implemented mocking for serial port connections in tests, removing the dependency on the physical `/dev/ttyACM0` device. This change enables tests to run successfully in both local and CI/CD environments. Additionally, configured automated testing on the server to ensure all tests are executed and verified on each pull request.